project/requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.27.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.89.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
minio==7.1.13
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
